Forgot to add the compiled version.
If you want a precompiled version, here it is! If you think there is a rat or a piece of malware in here, please use a decompiler on it: you will see that it is not obfuscated and that it clearly modifies your versions file for the specific Optifine version required for this to work.